test(multidb): consolidate 14 test files into 3
Fold the MultiDB tests into three files so the review diff is easier to read:

  multidb_test.go           package redis_test (unchanged)
  multidb_pipeline_test.go  package redis_test, renamed from
                            multidb_autopipeline_test.go (git mv, blame kept)
  multidb_internal_test.go  package redis, the twelve *_internal_test.go files
                            merged (imports unioned, bodies concatenated)

The split follows the only boundary Go enforces: internal (package redis) and
external (package redis_test) tests cannot share a file, so the internal
pipeline tests live in multidb_internal_test.go rather than the pipeline file.
No test was added, removed or changed: go test -list is identical before and
after (670 tests), and the multidb suites pass under -race.
